backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
prometheus-client
prometheus-fastapi-instrumentator

